Split lib.rs into modules
Closes https://github.com/PistonDevelopers/geometry/issues/16
