+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 
 name = "geometry"
-version = "0.0.7"
+version = "0.0.8"
 authors = [
     "bvssvni <bvssvni@gmail.com>"
 ]
@@ -24,4 +24,3 @@ git = "https://github.com/pistondevelopers/range"
 [dependencies.wavefront_obj]
 git = "https://github.com/PistonDevelopers/wavefront_obj"
 #version = "0.0.0"
-
