+++ src/model.rs
@@ -1,0 +1,16 @@
+use range::{ Range, ParentRange };
+
+use Object;
+
+/// A model consists of a list of object.
+/// The objects are stored separately,
+/// in an object buffer.
+pub struct Model(Range);
+
+impl ParentRange for Model {
+    type Child = Object;
+
+    fn from_range(range: Range) -> Self { Model(range) }
+    fn range(&self) -> &Range { &self.0 }
+    fn range_mut(&mut self) -> &mut Range { &mut self.0 }
+}
